getEvents query with all options besides match%
