+++ graphql/resolvers/index.js
@@ -1,6 +1,6 @@
 const usersResolver = require("./users");
 const eventsResolver = require("./events");
-
+const dateScalar = require("./date")
 module.exports = {
     Query: {
         ...usersResolver.Query,
@@ -10,5 +10,9 @@ module.exports = {
     Mutation: {
         ...usersResolver.Mutation,
         ...eventsResolver.Mutation,
+    },
+
+    Date: {
+        ...dateScalar.Date
     }
 };
